\docType{package}
\name{boussinesq-package}
\alias{boussinesq}
\alias{boussinesq-package}
\title{Analytic solutions for (ground-water) Boussinesq Equation}
\description{
  Analytic solutions for (ground-water) Boussinesq Equation
}
\details{
  \tabular{ll}{ Package: \tab boussinesq\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2011-12-06\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr
  Depends: \tab R(>=2.12)\cr }
}
\author{
  Emanuele Cordano \email{emanuele.cordano@gmail.org}
}
\keyword{package}

