\name{coefficient.song.solution}
\alias{coefficient.song.solution}
\title{Alogoritm for resolution of the series coefficient \eqn{a_n} for the dimensionless formula for \eqn{H} in \code{\link{beq.song.dimensionless}}}
\usage{
  coefficient.song.solution(n = 4, lambda = 0)
}
\arguments{
  \item{n}{approcimation order}

  \item{lambda}{dimensionless parameter releted to
  \eqn{\alpha} see Song at al, 2007}
}
\value{
  the \eqn{a_n} series coefficient
}
\description{
  Alogoritm for resolution of the series coefficient
  \eqn{a_n} for the dimensionless formula for \eqn{H} in
  \code{\link{beq.song.dimensionless}}
}
\note{
  For major details, see Song at al, 2007
}
\author{
  Emanuele Cordano
}
\references{
  Song, Zhi-yao;Li, Ling;David, Lockington. (2007), "Note
  on Barenblatt power series solution to Boussinesq
  equation",Applied Mathematics and Mechanics,
  \url{http://www.springerlink.com/content/w0u8667772712801/}
  ,\url{http://dx.doi.org/10.1007/s10483-007-0612-x}
}

