% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boussinesq-package.R
\docType{package}
\name{boussinesq-package}
\alias{boussinesq-package}
\alias{boussinesq}
\title{Analytic solutions for (ground-water) Boussinesq Equation}
\description{
Analytic solutions for (ground-water) Boussinesq Equation
}
\author{
Emanuele Cordano \email{emanuele.cordano@gmail.org}
}
\keyword{package}
