% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boussinesq-package.R
\docType{package}
\name{boussinesq-package}
\alias{boussinesq-package}
\alias{boussinesq}
\title{Analytic solutions for (ground-water) Boussinesq Equation}
\description{
Analytic solutions for (ground-water) Boussinesq Equation
}
\details{
\tabular{ll}{
Package: \tab boussinesq\cr
Type: \tab Package\cr
Version: \tab 1.0.2\cr
Date: \tab   2013-04-18\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
Depends: \tab R(>=2.12)\cr
}
}
\author{
Emanuele Cordano \email{emanuele.cordano@gmail.org}
}
\keyword{package}
