% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient.song.solution.R
\name{coefficient.song.solution}
\alias{coefficient.song.solution}
\title{Alogorithm for resolution of the series coefficient \eqn{a_n} for the dimensionless formula for \eqn{H} in \code{\link{beq.song.dimensionless}}}
\usage{
coefficient.song.solution(n = 4, lambda = 0)
}
\arguments{
\item{n}{approximation order}

\item{lambda}{dimensionless parameter related to \eqn{\alpha} see Song at al, 2007}
}
\value{
the \eqn{a_n} series coefficient
}
\description{
Alogorithm for resolution of the series coefficient \eqn{a_n} for the dimensionless formula for \eqn{H} in \code{\link{beq.song.dimensionless}}
}
\note{
For major details, see Song at al, 2007
}
\references{
Song, Zhi-yao;Li, Ling;David, Lockington. (2007), "Note on Barenblatt power series solution to Boussinesq equation",Applied Mathematics and Mechanics,
\url{https://link.springer.com/article/10.1007/s10483-007-0612-x} ,\doi{10.1007/s10483-007-0612-x}
}
\author{
Emanuele Cordano
}
